requirements.txt
flask
flask-cors
langchain-groq
langchain-core
langgraph
python-dotenv
gunicorn